Add most of the original grammar to the new PEG, add memoization
Add most of the original LALRPOP grammar to the PEG version, which was
tricky for the unary operator precedence climbing. Even the short test
program then took a noticeable amount of time to parse (infinite
lookahead parsers are N^2 by default). Added memoization to each
production based on offset, which was tricky because Rust makes things
like that tricky for thread safety. That and I'm stupid.